Refactor Telegram gateway into cohesive modules
