+++ packages/agent-telegram/src/Agent/Telegram/Internal/Runtime/Types.hs
@@ -1,0 +1,196 @@
+module Agent.Telegram.Internal.Runtime.Types (TelegramRuntime(..)) where
+
+
+import Agent.CLI.AgentSessions
+    ( SessionProcessLifetime(..)
+    , SessionProcessManager
+    , closeSessionProcessManager
+    , launchManagedTurnBounded
+    , newSessionProcessManagerWithLifetime
+    )
+import Agent.CLI.ManagedTurn
+    ( ManagedTurnMedia(..)
+    , ManagedTurnContext(..)
+    , ManagedTurnRequest(..)
+    , managedTurnRequestFromText
+    , managedTurnRequestWithGateway
+    )
+import Agent.CLI.Options
+    ( ApprovalPolicy(..)
+    , defaultEffortFor
+    )
+import Agent.CLI.Models
+    ( ModelOption(..)
+    , ModelTarget(..)
+    , defaultModelOptionFor
+    , rawModelOption
+    , resolveConfiguredModel
+    , resolveModelOptionDialect
+    )
+import Agent.CLI.ModelConfig (loadModelCatalog)
+import Agent.CLI.Session
+    ( SessionCreate(..)
+    , SessionHandle(..)
+    , SessionMeta(..)
+    , SessionTurn(..)
+    , SessionTurnPage(..)
+    , createSession
+    , loadSessionHandle
+    , loadRecentSessionTurns
+    , sessionTitleFromPrompt
+    , sessionsRoot
+    )
+import Agent.Telegram.Types
+import Agent.Telegram.Classify
+    ( TelegramUpdateAction(..)
+    , ambientGroupPrompt
+    , checkpointPendingVoiceTranscript
+    , classifyTelegramUpdate
+    , classifyTelegramUpdateWithMode
+    , grantableTelegramUser
+    , groupJoinAuthorized
+    , isAnonymousAdmin
+    , telegramAnonymousAdminUserId
+    , isAmbientGroupPrompt
+    , nextPendingAction
+    , reactionMessageText
+    , recordSeenTelegramUsers
+    , resolveTelegramUser
+    , storeUpdateAction
+    , telegramCommand
+    , telegramCommandArguments
+    , telegramReactionEmoji
+    , telegramReplyText
+    , telegramReplyUserIdFromPrompt
+    , telegramUserLabel
+    , TelegramUserResolution(..)
+    )
+import Agent.Telegram.Bridge
+    ( TelegramBridgeEnv(..)
+    , processTelegramCallbacks
+    , withTelegramBridge
+    )
+import qualified Agent.Telegram.Client as TelegramClient
+import Agent.Telegram.Log (logTelegramEvent)
+import Agent.Telegram.Markdown
+    ( markdownToTelegramHtml
+    , telegramRenderedLength
+    )
+import Agent.Telegram.Voice (transcribeWithXAI)
+import Agent.FileRetry (writeLazyFileAtomically)
+import Agent.Concurrent (mapConcurrentlyBounded)
+import Agent.OsPath (unsafeToFilePath)
+import Agent.Provider (Provider, parseProvider)
+import Agent.ReasoningEffort (reasoningEffortText)
+import Agent.Store.Postgres
+    ( Store
+    , managedPostgresConfigFromEnv
+    , trustedPool
+    , withStore
+    )
+import Agent.Store.Postgres.Connection (StorePool)
+import Agent.Store.Types (renderStoreError)
+import Control.Concurrent (threadDelay)
+import Control.Concurrent.Async
+    ( replicateConcurrently_
+    , race_
+    , withAsync
+    )
+import Control.Concurrent.Chan (Chan, newChan, readChan, writeChan)
+import Control.Concurrent.MVar
+    ( MVar
+    , modifyMVar
+    , modifyMVar_
+    , newMVar
+    , readMVar
+    )
+import Control.Exception.Safe
+    ( SomeException
+    , bracket
+    , displayException
+    , finally
+    , onException
+    , try
+    , tryAny
+    )
+import Control.Monad (forM_, unless, void, when)
+import Data.Aeson
+    ( Value(..)
+    , eitherDecode
+    , encode
+    , object
+    , (.=)
+    )
+import qualified Data.ByteString.Lazy as LBS
+import Data.Int (Int64)
+import qualified Data.Text.Encoding as TextEncoding
+import qualified Data.Map.Strict as Map
+import Data.List (sortOn)
+import Data.Maybe (fromMaybe)
+import Data.Set (Set)
+import qualified Data.Set as Set
+import Data.Text (Text)
+import qualified Data.Text as Text
+import qualified Data.Text.IO as Text
+import Data.Time.Clock
+    ( addUTCTime
+    , diffUTCTime
+    , getCurrentTime
+    )
+import qualified Network.HTTP.Client.TLS as HttpTls
+import qualified System.Directory as Directory
+import System.Directory.OsPath
+    ( createDirectoryIfMissing
+    , doesFileExist
+    , getCurrentDirectory
+    , getHomeDirectory
+    , makeAbsolute
+    , removeFile
+    )
+import System.Environment (getArgs, getExecutablePath, lookupEnv)
+import System.Exit (die)
+import System.FilePath (takeExtension)
+import System.IO
+    ( IOMode(AppendMode)
+    , hFlush
+    , hGetEcho
+    , hIsTerminalDevice
+    , hSetEcho
+    , withFile
+    , stderr
+    , stdin
+    )
+import System.OsPath (OsPath, unsafeEncodeUtf, (</>))
+import System.Posix.Files (setFileMode)
+import System.Posix.Process (getProcessID)
+import System.Posix.Signals (nullSignal, sigTERM, signalProcess)
+import System.Posix.Types (ProcessID)
+import System.Process
+    ( CreateProcess(..)
+    , ProcessHandle
+    , StdStream(..)
+    , createProcess
+    , getProcessExitCode
+    , proc
+    )
+import qualified System.FileLock as FileLock
+import Text.Read (readMaybe)
+
+data TelegramRuntime = TelegramRuntime
+    { runtimeClient :: !TelegramClient
+    , runtimeBot :: !TelegramUser
+    , runtimeRespondToAllGroupMessages :: !Bool
+    , runtimeWorkerCount :: !Int
+    , runtimeGatewayDirectory :: !OsPath
+    , runtimePool :: !StorePool
+    , runtimeSessionsRoot :: !OsPath
+    , runtimeStatePath :: !OsPath
+    , runtimeStateVar :: !(MVar TelegramState)
+    , runtimeWorkQueue :: !(Chan TelegramChatKey)
+    , runtimeScheduled :: !(MVar (Set TelegramChatKey))
+    , runtimeProcessManager :: !SessionProcessManager
+    , runtimeTarget :: !ModelTarget
+    , runtimeCwd :: !OsPath
+    , runtimeEffort :: !Text
+    , runtimePolicy :: !ApprovalPolicy
+    }
